fix(bond): resolve all pre-existing clippy -D warnings
63 warnings that existed before the fork consolidation now pass
-D warnings. Changes are non-behavioral:

- Remove unused `String` import from slashing.rs
- Prefix unused `key` param with `_key` in nonce::bump_nonce_ttl
- Change inner `///` doc comments (invalid in fn bodies) to `//` in lib.rs
- Replace `u32::try_from(list_len).unwrap_or(u32::MAX)` with `list_len`
  in invariants.rs (Soroban Vec::len already returns u32)
- Remove `as u32` no-op cast in types/attestation.rs
- Replace `unwrap_or_else(|| ...)` with `unwrap_or(...)` in tiered_bond.rs
- Add `#![allow(dead_code)]` to batch.rs, claims.rs, upgrade_auth.rs
  (incomplete feature modules not yet wired to the contract)
- Add targeted `#[allow(dead_code)]` to unused-but-intentional helpers
  in events.rs, nonce.rs, rolling_bond.rs, same_ledger_liquidation_guard.rs,
  slashing.rs

cargo clippy -p credence_bond -- -D warnings: clean (0 errors)
cargo test --workspace (excl. credence_admin_cli): all pass

+++ contracts/credence_bond/src/types/attestation.rs
@@ -66,7 +66,7 @@ impl Attestation {
     #[inline]
     pub fn validate_data(data: &String) {
         let len = data.len();
-        if len as u32 > MAX_ATTESTATION_DATA_LENGTH {
+        if len > MAX_ATTESTATION_DATA_LENGTH {
             panic!("attestation data exceeds maximum length");
         }
     }
